Made electric-fence an optional dependency
On CentOS-7 with electric-fence calls to malloc() were deadlocking in some
dependent packages we were building. Using electric-fence is not
necessarily required for development, it's just a development tool, so
we want to make it conditional.

OpenSSL allows disabling it by using the debug-linux-x86_64-noefence
configuration template, but that is abstracted away and not directly
accessible when using the config script.